+++ app/src/main/kotlin/ktx/app/graphics.kt
@@ -1,0 +1,24 @@
+package ktx.app
+
+import com.badlogic.gdx.graphics.g2d.Batch
+import com.badlogic.gdx.graphics.glutils.ShaderProgram
+
+/**
+ * Automatically calls [Batch.begin] and [Batch.end].
+ * @param action inlined. Executed after [Batch.begin] and before [Batch.end].
+ */
+inline fun <B : Batch> B.use(action: (B) -> Unit) {
+  begin()
+  action(this)
+  end()
+}
+
+/**
+ * Automatically calls [ShaderProgram.begin] and [ShaderProgram.end].
+ * @param action inlined. Executed after [ShaderProgram.begin] and before [ShaderProgram.end].
+ */
+inline fun <S : ShaderProgram> S.use(action: (S) -> Unit) {
+  begin()
+  action(this)
+  end()
+}